Add a filter to allow for changing the format of the dropdown boxes
I would prefer to have the dropdown box use the topic post_title rather than post_content, which is completely uninformative in my opinion. Regardless of whether you agree/disagree, if you add a filter here, then people can easily customize it to their liking. 

An alternative is making it a pluggable function with "if ! function_exists ( )" wrapping it.