add values to defaults array
these parameters are passed in from bp_document_get(), and then used within the BP_Document::get() function. However, they are not listed in the default array. Functionally this might not matter much, but it was very difficult to figure out which arguments I could actually use with bp_document_get().