Fix Entry Bubble Buttons redundancies
There's an immense amount of redundant code in the generation of the Activity Entry Bubble Buttons. The Edit and and Delete buttons are fully generated for ALL activity posts, and then removed when the user doesn't have permission to do so (which would surely be the VAST majority of the time, given that most posts are not made by a single user). Also, the `bp_get_activity_delete_url()` function is called twice in succession, which is needless.